Module Stdppx.Char

include module type of struct include Char end

Characters

type t = char

An alias for the type of characters.

val code : char -> int

Return the integer code of the argument.

val chr : int -> char

Return the character with the given integer code.

raises Invalid_argument
if the argument is outside the range [0x00;0xFF].

val escaped : char -> string

Return a string representing the given character, with special characters escaped following the lexical conventions of OCaml. All characters outside the ASCII printable range [0x20;0x7E] are escaped, as well as backslash, double-quote, and single-quote.

Predicates and comparisons

See also the Ascii module.

ASCII characters

module Ascii = Char.Ascii

ASCII character set support.

val lowercase_ascii : char -> char

Use the equivalent Ascii.lowercase instead.

val uppercase_ascii : char -> char

Use the equivalent Ascii.uppercase instead.

Hashing

val seeded_hash : int -> t -> int

A seeded hash function for characters, with the same output value as Hashtbl.seeded_hash. This function allows this module to be passed as argument to the functor Hashtbl.MakeSeeded.

val hash : t -> int

An unseeded hash function for characters, with the same output value as Hashtbl.hash. This function allows this module to be passed as argument to the functor Hashtbl.Make.

include Comparisons with type t := char

val compare : char -> char -> int

val equal : char -> char -> bool

val (=) : char -> char -> bool

val (<) : char -> char -> bool

val (>) : char -> char -> bool

val (<>) : char -> char -> bool

val (<=) : char -> char -> bool

val (>=) : char -> char -> bool

val min : char -> char -> char

val max : char -> char -> char